// src/types/book-detail.ts
export type BookWork = {
  description: string;
  title: string;
  covers: number[];
  subject_places: string[];
  subject_times: string[];
  subjects: string[];
};

export type BookDataDetail = {
  description: string;
  title: string;
  image: string;
  subjectPlaces: string;
  subjectTimes: string;
  subjects: string;
};